Clean up from concurrency merges
